Replace `ensure_resource` with base class
By having a base class that can be included by multiple instances of
`spiped::tunnel`, we can remove the use of `ensure_resource`.

This also makes it possible to provide parameters via the base class.
I've added `package_source`.  This parameter can be used on OSes that
don't ship a `spiped` package.

This commit also contains new rspec-puppet tests and a README update
explaining how to use the module on RedHat family systems.